--- a/Neos.ContentGraph.PostgreSQLAdapter/src/Domain/Projection/HypergraphProjection.php
+++ b/Neos.ContentGraph.PostgreSQLAdapter/src/Domain/Projection/HypergraphProjection.php
@@ -219,9 +219,11 @@ final class HypergraphProjection implements ContentGraphProjectionInterface
             throw new \RuntimeException(sprintf('Invoking %s is not allowed to be invoked recursively. Current transaction nesting %d.', __FUNCTION__, $this->dbal->getTransactionNestingLevel()));
         }
         $this->dbal->beginTransaction();
+        $this->dbal->setRollbackOnly();
         try {
             return $fn();
         } finally {
+            // unsets rollback only flag and allows the connection to work regular again
             $this->dbal->rollBack();
         }
     }