--- a/Neos.ContentRepository.Core/Classes/CommandHandler/CommandBus.php
+++ b/Neos.ContentRepository.Core/Classes/CommandHandler/CommandBus.php
@@ -28,7 +28,7 @@ final class CommandBus
     }
 
     /**
-     * @return EventsToPublish|\Generator<int, EventsToPublish, ?EventsToPublishFailed, void>
+     * @return EventsToPublish|\Generator<int, EventsToPublish>
      */
     public function handle(CommandInterface $command, CommandHandlingDependencies $commandHandlingDependencies): EventsToPublish|\Generator
     {