--- a/Neos.ContentRepository.Core/Classes/CommandHandler/CommandBus.php
+++ b/Neos.ContentRepository.Core/Classes/CommandHandler/CommandBus.php
@@ -7,7 +7,6 @@ namespace Neos\ContentRepository\Core\CommandHandler;
 use Neos\ContentRepository\Core\CommandHandlingDependencies;
 use Neos\ContentRepository\Core\ContentRepository;
 use Neos\ContentRepository\Core\EventStore\EventsToPublish;
-use Neos\ContentRepository\Core\EventStore\EventsToPublishFailed;
 
 /**
  * Implementation Detail of {@see ContentRepository::handle}, which does the command dispatching to the different