--- a/Neos.ContentRepository.Core/Classes/CommandHandler/CommandHandlerInterface.php
+++ b/Neos.ContentRepository.Core/Classes/CommandHandler/CommandHandlerInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Neos\ContentRepository\Core\CommandHandler;
 
-use Neos\ContentRepository\Core\CommandHandlingDependencies;
 use Neos\ContentRepository\Core\EventStore\EventsToPublish;
 
 /**