--- a/Neos.ContentRepository.Core/Classes/CommandHandler/CommandHandlerInterface.php
+++ b/Neos.ContentRepository.Core/Classes/CommandHandler/CommandHandlerInterface.php
@@ -20,7 +20,7 @@ interface CommandHandlerInterface
     /**
      * "simple" command handlers return EventsToPublish directly
      *
-     * For the case of the workspace command handler who need to publish to many streams and "close" the content-stream directly,
+     * For the case of the workspace command handler that need to publish to many streams and "close" the content-stream directly,
      * it's allowed to yield the events to interact with the control flow of event publishing.
      *
      * @return EventsToPublish|\Generator<int, EventsToPublish>