--- a/Neos.ContentRepository.Core/Classes/Projection/ProjectionInterface.php
+++ b/Neos.ContentRepository.Core/Classes/Projection/ProjectionInterface.php
@@ -47,4 +47,6 @@ interface ProjectionInterface
     public function getState(): ProjectionStateInterface;
 
     public function reset(): void;
+
+    // public function temporaryChanges(\Closure $fn): void;
 }