@kitsunet 2024-10-23
Do we see any risks in that the _one_ connection used here is shared between ORM and eventstore and the CR? Given we _must_ ensure nothing commits the transaction we started for the simulation. So no hooks, signals or anything that might lead to userland code being executed that could in some way start/commit a transaction.