@kitsunet 2024-10-25
Alright, at least according to the tests this works now. I also went through all Rebasable commands to check if the events get enriched. the Dimension ones were the only missing. IMHO we _should_ centralize this behavior, I opted against doing it here though as I think we need to consider if there would have to be any more logic involved to decide what gets enriched with commands, in what way, when we override the command if there is already metadata and finally what the causation ids are. I _guess_ we could ignore all these questions and centralize it, but it warrants a closer look and is therefore out of scope of this change.